// Avocat_Angular-master/src/app/models/type/contributorType.ts
export enum ContributorType {
    JUDGE = 'JUDGE',
    JUDICIAL_OFFICER = 'JUDICIAL_OFFICER',
    LEGAL_EXPERT = 'LEGAL_EXPERT',
    DEFENDANT = 'DEFENDANT',
    PLAIGNANT = 'PLAIGNANT',
    WITNESS = 'WITNESS',
    REPORTER = 'REPORTER',
    LEGAL_COUNSEL = 'LEGAL_COUNSEL',
    INVESTIGATING_JUDGE = 'INVESTIGATING_JUDGE',
    CUSTOMER = 'CUSTOMER',
    LAWYER = 'LAWYER'
  }
  